Add files via upload
Dataset Explanation: 
The stroke dataset is a collection of data related to individuals who have had a stroke, as well as some who have not. It contains information on various factors that are known to be associated with stroke, including age, gender, marital status, hypertension, heart disease, occupation, place of residence, smoking status, and various health metrics. The dataset is commonly used in machine learning and statistical analysis to develop predictive models for stroke risk.

Here is a brief explanation of each column in the dataset:

id: A unique identifier for everyone in the dataset.
gender: The gender of the individual, recorded as "Male", "Female", or "Other".
age: The age of the individual in years.
married: Whether the individual is currently married or not, recorded as "Yes" or "No".
hypertension: Whether the individual has hypertension (high blood pressure), recorded as 1 if present and 0 if not.
heart_disease: Whether the individual has a history of heart disease, recorded as 1 if present and 0 if not.
occupation: The occupation of the individual, recorded as "Goverment Job", "Private Job", "Self-employed", "Children", or "Never Worked".
residence: The type of residence of the individual, recorded as "Urban" or "Rural".
metric_1: A health metric related to the level of glucose in the blood.
metric_2: A health metric related to the level of body mass index (BMI).
metric_3: A health metric related to the level of physical activity.
metric_4: A health metric related to the level of total cholesterol in the blood.
metric_5: A health metric related to the level of systolic blood pressure.
smoking_status: The smoking status of the individual, recorded as "formerly smoked", "never smoked", "smokes", or "Unknown".
stroke: Whether the individual has had a stroke, recorded as 1 if yes and 0 if no.
Overall, this dataset provides a rich set of information about individuals who have had strokes, which can be used to develop models for predicting stroke risk and improving stroke prevention efforts.